Remove static MP event and replace with instances of category

+++ src/RageMP.Net/Enums/EntityType.cs
@@ -1,6 +1,6 @@
 namespace RageMP.Net.Enums
 {
-    public enum EntityType : ushort
+    public enum EntityType : byte
     {
         Player,
         Vehicle,
